`@repo/mui` internal package extraction
