+++ apps/emergence/app/about/page.tsx
@@ -4,8 +4,8 @@ import Typography from '@mui/material/Typography'
 import Box from '@mui/material/Box'
 import Button from '@mui/material/Button'
 import NextLink from 'next/link'
-import ProTip from '@/components/ProTip'
-import Copyright from '@/components/Copyright'
+import ProTip from '@repo/mui/components/ProTip'
+import Copyright from '@repo/mui/components/Copyright'
 
 export default function About() {
   return (
