+++ apps/emergence/app/favicon.ico
