+++ apps/emergence/app/layout.tsx
@@ -3,8 +3,8 @@ import { AppRouterCacheProvider } from '@mui/material-nextjs/v15-appRouter'
 import { ThemeProvider } from '@mui/material/styles'
 import CssBaseline from '@mui/material/CssBaseline'
 import InitColorSchemeScript from '@mui/material/InitColorSchemeScript'
-import theme from '@/theme'
-import ModeSwitch from '@/components/ModeSwitch'
+import theme from '@repo/mui/theme'
+import ModeSwitch from '@repo/mui/components/ModeSwitch'
 
 export default function RootLayout(props: { children: React.ReactNode }) {
   return (
