+++ apps/emergence/app/page.tsx
@@ -4,8 +4,8 @@ import Typography from '@mui/material/Typography'
 import Box from '@mui/material/Box'
 import Link from '@mui/material/Link'
 import NextLink from 'next/link'
-import ProTip from '@/components/ProTip'
-import Copyright from '@/components/Copyright'
+import Copyright from '@repo/mui/components/Copyright'
+import ProTip from '@repo/mui/components/ProTip'
 
 export default function Home() {
   return (
