+++ packages/mui/eslint.config.ts
@@ -1,0 +1,6 @@
+import { reactConfig } from '@repo/eslint-config/react'
+import { type ConfigArray } from 'typescript-eslint'
+
+const config: ConfigArray = reactConfig(true)
+
+export default config
