+++ packages/mui/package.json
@@ -1,14 +1,16 @@
 {
-  "name": "emergence",
+  "name": "@repo/mui",
   "private": true,
   "version": "0.0.0",
+  "exports": {
+    "./*": "./src/*.ts",
+    "./components/*": "./src/components/*.tsx"
+  },
   "scripts": {
-    "build": "next build",
     "check-types": "tsc --noEmit",
-    "dev": "next dev --port 9000",
-    "format": "next lint --fix --max-warnings 0",
-    "lint": "next lint --max-warnings 0",
-    "start": "next start"
+    "format": "eslint . --fix --max-warnings 0",
+    "generate:component": "turbo gen react-component",
+    "lint": "eslint . --max-warnings 0"
   },
   "dependencies": {
     "@emotion/cache": "^11.14.0",
@@ -17,24 +19,17 @@
     "@mui/icons-material": "^7.2.0",
     "@mui/material": "^7.2.0",
     "@mui/material-nextjs": "^7.2.0",
-    "@repo/mui": "workspace:*",
     "next": "^15.3.5",
     "react": "^19.1.0",
     "react-dom": "^19.1.0"
   },
   "devDependencies": {
-    "@eslint/compat": "^1.3.1",
-    "@eslint/eslintrc": "^3.3.1",
     "@repo/eslint-config": "workspace:*",
     "@repo/typescript-config": "workspace:*",
     "@types/node": "^24.0.14",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
     "eslint": "^9.31.0",
-    "eslint-config-next": "^15.4.1",
-    "eslint-config-prettier": "^10.1.8",
-    "eslint-plugin-prettier": "^5.5.3",
-    "jiti": "^2.4.2",
     "typescript": "^5.8.3"
   }
 }
