+++ packages/mui/src/components/ProTip.tsx
